提交代码

+++ src/main/java/com/pkusoft/agxt/model/FileTemp.java
@@ -83,7 +83,7 @@ public class FileTemp {
      */
     @Column(name = "SN")
     @ApiModelProperty(value = "排序序号")
-    private Integer sn;
+    private Double sn;
 
  	/**
      * 级别
@@ -416,7 +416,7 @@ public class FileTemp {
     * 获得排序序号
     * @return SN
     */
-    public Integer getSn() {
+    public Double getSn() {
         return this.sn;
     }
 
@@ -424,7 +424,7 @@ public class FileTemp {
     * 设置排序序号
     * @param sn
     */
-    public void setSn(Integer sn) {
+    public void setSn(Double sn) {
         this.sn = sn;
     }
